feat: Branch 관련 Facade 패턴으로 변환 - 추후, BranchService 는, ProjectService 등에서 재활용 될 예정
