+++ src/main/java/baekgwa/suhoserver/model/branch/type/repository/BranchTypeRepository.java
@@ -2,7 +2,6 @@ package baekgwa.suhoserver.model.branch.type.repository;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
@@ -30,8 +29,13 @@ public interface BranchTypeRepository extends JpaRepository<BranchTypeEntity, Lo
 	List<BranchTypeEntity> findLatest(
 		@Param("versionInfoId") Long versionInfoId,
 		@Param("code") String code,
-		Pageable pageable);
+		Pageable pageable
+	);
 
-	@Query("select b from BranchTypeEntity b where b.versionInfoEntity.id = :versionInfoId and b.code = :code and b.version = :version")
-	Optional<BranchTypeEntity> findBranchType(@Param("versionInfoId") Long versionInfoId, @Param("code") String code, @Param("version") LocalDate version);
+	@Query("SELECT (COUNT(b) > 0) FROM BranchTypeEntity b WHERE b.versionInfoEntity.id = :versionInfoId AND b.code = :code AND b.version = :versionDate")
+	boolean existsByParams(
+		@Param("versionInfoId") Long versionInfoId,
+		@Param("code") String code,
+		@Param("versionDate") LocalDate versionDate
+	);
 }
